Fix security vulnerability: Run Docker container as non-root user
Addresses SonarCloud security hotspot by adding a dedicated non-root
user to the Docker container. The container now runs as the 'sweeper'
user instead of root, following the principle of least privilege.

This reduces security risk if the container or application is
compromised, as attackers would not have root privileges within
the container.

Fixes #192